requirements.txt
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.0.76
mediapipe==0.10.5
flask==2.3.3
flask-sock>=0.6.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
pyttsx3==2.90
google-generativeai==0.3.0
requests==2.31.0 